pyproject.toml
[tool.poetry]
name = "leetcode-randomizer"
version = "0.1.0"
description = "Returns five random LC problems of selected pattern"
authors = ["rdturbo <rupayan98@gmail.com>"]
license = "MIT"
readme = "README.md"
homepage = "https://github.com/rdturbo/leetcode-randomizer"
repository = "https://github.com/rdturbo/leetcode-randomizer"
keywords = ["leetcode"]

[tool.poetry.dependencies]
python = "^3.10"
requests = "^2.28.1"
click = "^8.1.3"
python-dotenv = "^0.20.0"
python-dateutil = "^2.8.2"

[tool.poetry.dev-dependencies]
flake8 = "^5.0.4"
black = {version = "^22.8.0", allow-prereleases = true}
pytest = "^7.1.2"

[tool.poetry.scripts]
leetcode-randomizer = "src.leetcode_randomizer.console:main"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
